- Removed unneeded scan of uk.ac.ebi.eva.lib.dgva_utils - assertTrue >= 1 to assertEquals 3 - Moved eva-lib metadata tests to same package as classes they're testing
