+++ eva-lib/src/test/java/uk/ac/ebi/eva/lib/metadata/FileTestData.java
@@ -19,19 +19,19 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 
 import uk.ac.ebi.eva.lib.entities.File;
 
-final class FileTestData {
+public final class FileTestData {
 
-    static final String FILE_1_NAME = "file1.vcf.gz";
+    public static final String FILE_1_NAME = "file1.vcf.gz";
 
-    static final String FILE_2_NAME = "file2.vcf.gz";
+    public static final String FILE_2_NAME = "file2.vcf.gz";
 
-    static final String FILE_2_TABIX_NAME = FILE_2_NAME + ".tbi";
+    public static final String FILE_2_TABIX_NAME = FILE_2_NAME + ".tbi";
 
-    static final String FILE_NOT_BROWSABLE = "file.notBroswable.vcf.gz";
+    public static final String FILE_NOT_BROWSABLE = "file.notBroswable.vcf.gz";
 
     private FileTestData() {}
 
-    static void persistTestData(TestEntityManager entityManager) {
+    public static void persistTestData(TestEntityManager entityManager) {
         File file1 = new File(1L, "ERF1", FILE_1_NAME, "sd3245as8dasiu2345d", "/dir/path", "vcf", "submitted", 1, true,
                               "/parentdir/dir1/" + FILE_1_NAME, true, "EVAF1");
         File file2 = new File(2L, "ERF2", FILE_2_NAME, "zd32452343242345c", "/dir/path", "vcf", "submitted", 1, true,
